Add country detail page and features
Implemented a country detail page with routing and features. This included updating the tabs.page.html and app-routing.module.ts to accommodate the new page. A country model and services were also created to handle the fetching of country data from an API. Changes also made to tab1 to display list of countries.