Add README.md file to the project
A README.md file was added to give a thorough explanation of the project. It includes sections detailing how to run the project locally, features of the applications, tech stacks used, and lessons learned from the project. The new file will make it easier for others to understand and contribute to the project.