Extends Country model with MapCountry model
Code is modified to replace the Country model with the MapCountry model in the Tab1Page and CountryDetailPage components, mapping the Country model in the CountriesService. The model's change refactors the data representation to better handle the currency details, which are now more presentable in the view layer (CountryDetailPage), with the inherent JSON representation removed.