Enhance Tab1Page with modal, country search, and currency filter
The Tab1Page has been heavily extended with new functionalities and model references for a better user experience. These include the addition of an ion modal, a search function to filter through countries, and a feature to filter by currencies. The MapCurrency model is also now imported alongside MapCountry and the control flow for initialising countries is updated. These changes aim to make the user's navigation and the data presentation on the Tab1Page more efficient and enjoyable.